[skip ci] WIP: Generic adapters for warnings
* WIP
